Add files via upload
This Project involves doing analysis of sales data of a company called "BikeHaven". the project includes creation of a Sales Dashboard for the company as well as a workbook for the company to get valuable insights on sales such as top 10 customers, top 10 products, region wise sales, previous two year's sales volume, relation between sale amount and budget, relation between category of product and sales etc.. The entire project is done with the help of Tableu.